Refactoring: separate view and controller of RatesModule
