+++ Curventer/PresentationLayer/RatesModule/RatesVC.swift
@@ -1,14 +1,9 @@
 import UIKit
 
 class RatesVC: UIViewController {
-    // MARK: - Properties
-    let ratesView = RatesView()
-    let dataSource = RatesDataSource()
-    
     // MARK: - View lifecycle methods
     override func loadView() {
-        ratesView.tableView.dataSource = dataSource
-        ratesView.tableView.delegate = self
+        let ratesView = RatesView()
         self.view = ratesView
     }
     
@@ -26,34 +21,15 @@ class RatesVC: UIViewController {
         }
     }
     
-    private func removeCurrency(at index: Int) {
-        dataSource.rates.remove(at: index)
-    }
-    
     private func fetchRates(for base: String) {
         RateAPIManager.fetch(type: Rate.self, router: RateRouter.fetchLatestRatesOnBase(base)) { result in
             switch result {
             case .success(let rate):
-                self.dataSource.rates = self.prepareForDisplay(from: rate.rates)
-                self.ratesView.updateTable()
+                let rates = self.prepareForDisplay(from: rate.rates)
+                (self.view as? RatesView)?.updateTable(with: rates)
             case .failure(let error):
                 print("ERROR HANDLER: \(error.localizedDescription)")
             }
-        }
-    }
-}
-
-
-// MARK: - TableViewDelegates
-extension RatesVC: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-    
-    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete {
-            removeCurrency(at: indexPath.row)
-            tableView.deleteRows(at: [indexPath], with: .fade)
         }
     }
 }
